requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==4.2.23
django-ranged-response==0.2.0
django-rest-framework==0.1.0
django-simple-captcha==0.6.2
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
numpy==2.0.2
packaging==25.0
pillow==11.3.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
